% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCrossingScenario.R
\name{AddRoadUserCornerMatrices}
\alias{AddRoadUserCornerMatrices}
\title{AddRoadUserCornerMatrices.m}
\usage{
AddRoadUserCornerMatrices(x, b, type = "Approaching")
}
\arguments{
\item{x}{a Pedestrian or a Driver object}

\item{b}{a Behaviour object}

\item{type}{either Approaching or Crossing behaviour}
}
\description{
add a field MCornerMatrices, describing the locations of the four corners 
of the road user over time in the scenario, to the input/output structure
SRoadUserBehaviour, based on the VX, VY, and VHeading fields of that 
structure, and the fields length and width of the SRoadUserConstants 
input structure
}
