% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{analyse_interactome}
\alias{analyse_interactome}
\title{Construct an interactome by comparing bait and control background across experimental conditions}
\usage{
analyse_interactome(df, Intensity, ibait, bait_gene_name, Npep, Protein.IDs,
  name_bait, name_ctrl, background, conditions, replicates,
  by_conditions = TRUE, pool_background = TRUE, log_test = TRUE,
  log_stoichio = TRUE, substract_ctrl = TRUE, use_mean_for_bait = TRUE)
}
\arguments{
\item{df}{a data frame of protein intensities. columns are experimental samples and rows are proteins}

\item{Intensity}{a data frame of protein intensities (without replacement of missing values). columns are experimental samples and rows are proteins}

\item{ibait}{: row index corresponding to the bait protein}

\item{bait_gene_name}{: The gene name of the bait}

\item{Npep}{: vector containing the number of theoretically observable peptide per protein (same length as \code{dim(df)[1]})}

\item{Protein.IDs}{: vector containing protein IDs (same length as \code{dim(df)[1]})}

\item{name_bait}{: name of the bait as appearing in the background vector}

\item{name_ctrl}{: name of the control as appearing in the background vector}

\item{background}{: vector of background names for each experimental sample}

\item{conditions}{: vector of conditions for each each experimental sample}

\item{replicates}{: vector of biological replicates for each each experimental sample}

\item{by_conditions}{option to perform the comparison between bait and control group for each condition}

\item{pool_background}{option to use all control background conditions as one control group for all conditions}

\item{log_test}{logical, perform t-test on log transform intensities}

\item{log_stoichio}{logical, use the geometric mean instead of the arithmetic mean to compute stoichiometries}

\item{substract_ctrl}{logical, substract ctrl intensities in the calculation of stoichiometries}
}
\value{
an object of class \code{InteRactome}, i.e a list including the following elements :

\code{conditions} : a vector of experimental conditions.

\code{names} : a vector of names (by default gene names are used).

\code{p_val} : a list of vectors containing the p values associated to each experimental condition.

\code{fold_change} : a list of vectors containing the fold change associated to each experimental condition.

\code{...} : other variables.
}
\description{
Construct an interactome by comparing bait and control background across experimental conditions
}
