% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{annotation_enrichment_analysis}
\alias{annotation_enrichment_analysis}
\title{Perform enrichment analysis}
\usage{
annotation_enrichment_analysis(df, idx_detect,
  annotation_selected = c("Keywords", "Protein.families"),
  names = df$Gene.names...primary.., organism = "mouse",
  updateProgress = NULL, showProgress = TRUE, orderOutput = TRUE)
}
\arguments{
\item{df}{a formatted data.frame with annoations corresponding to each row. Types of annotations are organized by columns.}

\item{names}{row names. Used in the output data.frame}

\item{organism}{organism for which the analysis is to be performed ("mouse" or "human")}

\item{updateProgress}{logical, function to show progress in shiny app}

\item{showProgress}{logical, show progress in console}

\item{orderOutput}{logical, order annotations by enrichment p-values in the output data.frame}

\item{idx_select}{indexes of the foreground set.}
}
\value{
a data.frame
}
\description{
Perform enrichment analysis for protein annotations stored in a formatted data.frame
}
