% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{append_FDR.InteRactome}
\alias{append_FDR.InteRactome}
\title{Append a FDR column to an \code{InteRactome}}
\usage{
\method{append_FDR}{InteRactome}(res, df)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{df}{a data.frame containing (at least) columns 'bait', names', 'FDR' and 'conditions'}
}
\value{
an \code{InteRactome}
}
\description{
Append a FDR column to an \code{InteRactome}
}
\examples{
df_merge <- merge_conditions(Interactome)
df_FDR <- compute_FDR_from_asymmetry(df_merge)
Interactome <- append_FDR(Interactome, df_FDR)
}
