% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{append_FDR}
\alias{append_FDR}
\title{Append a FDR column to an \code{InteRactome}}
\usage{
append_FDR(res, df)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{df}{a data.frame containing (at least) columns 'id', 'bait', 
'names', 'FDR' and 'conditions'}
}
\value{
an \code{InteRactome}
}
\description{
Append a FDR column to an \code{InteRactome}
}
\examples{
#' #load data :
data("proteinGroups_Cbl")
#Run InteRact with default parameters
res <- InteRact(proteinGroups_Cbl, bait_gene_name = "Cbl")
df_merge <- merge_conditions(res)
df_FDR <- compute_FDR_from_asymmetry(df_merge)
Interactome <- append_FDR(res, cbind(df_merge, FDR = df_FDR$FDR) )
}
