% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{append_PPI}
\alias{append_PPI}
\title{Append protein-protein interaction}
\usage{
append_PPI(res, mapping = "names")
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{mapping}{name of the \code{InteRactome}'s variable containing gene names}
}
\value{
an \code{InteRactome}
}
\description{
Append protein-protein interaction information to an \code{InteRactome}.
PPI are retrieved from databases IntAct, MINT, BioGRID and HPRD
}
