% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{average_technical_replicates}
\alias{average_technical_replicates}
\title{Averaging of protein intensities over technical replicates}
\usage{
average_technical_replicates(df, cond)
}
\arguments{
\item{df}{A data frame of protein intensities}

\item{conditions}{A data frame containing the description of df's columns (i.e "idx", bckg", "time", "bio"  and "tech")}
}
\value{
A list containing :
\code{Intensity}, a data frame of protein intensities averaged over technical replicates;
\code{conditions}, a data frame containing the description of Intensity's columns
}
\description{
Averaging of protein intensities over technical replicates
}
