% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_characteristic_emotion.R
\name{batch_characteristic_emotion}
\alias{batch_characteristic_emotion}
\title{Batch Characteristic Emotion}
\usage{
batch_characteristic_emotion(df)
}
\arguments{
\item{df}{dataframe with a variable called term that corresponds to identities you want characteristic emotions for}
}
\value{
df with 3 new variables: characteristic emotion E, P, and A
}
\description{
Batch Characteristic Emotion
}
