% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{Behaviour-class}
\alias{Behaviour-class}
\title{An S4 class of Approaching Road User Behaviour.}
\description{
Details of the class...
}
\section{Slots}{

\describe{
\item{\code{VY}}{a vector. Travelling along the X axis}

\item{\code{VX}}{a vector. Get position signal}

\item{\code{VXdot}}{a vector. -VSpeed. Save also the speed signal for later use}

\item{\code{VHeading}}{a vector. Always heading West}

\item{\code{VbYieldingeHMI}}{a logical vector. From eHMI section.}
}}

