% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_emotion.R
\name{characteristic_emotion}
\alias{characteristic_emotion}
\title{Characteristic Emotion}
\usage{
characteristic_emotion(id)
}
\arguments{
\item{id}{lowercase string corresponding to the identity}
}
\value{
3 digit EPA profile of characteristic emotion for the identity
}
\description{
Characteristic Emotion
}
\examples{
characteristic_emotion("brute")
}
