% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_h_matrix.R
\name{construct_h_matrix}
\alias{construct_h_matrix}
\title{Construct H Matrix}
\usage{
construct_h_matrix(eq)
}
\arguments{
\item{eq}{the equation dataframe being used in the analysis}
}
\value{
the h matrix, which "symbolize the matrices of constant parameters"
(see Heise, Expressive Order, p. 87)
}
\description{
This function is for internal use within other functions. A helper function to
make the H matrix needed in ACT calculations
}
\examples{
eq <- get_equation(name = "us2010", type = "impressionabo", g = "average")
h_mat <- construct_h_matrix(eq = eq)
}
