% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{discretize_values}
\alias{discretize_values}
\title{Discretize values in a vector based on a finite set of values}
\usage{
discretize_values(x, breaks = c(1, 0.1, 0.05, 0.01),
  decreasing_order = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{breaks}{numeric vector. Set of discrete values on which \code{x} values will be mapped. 
Non-mapped values will be set to NA}

\item{decreasing_order}{logical. Map \code{beaks} values from the greatest to the smallest}
}
\value{
a numeric vector
}
\description{
Discretize values in a vector based on a finite set of values
}
