% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{estimate_Npep}
\alias{estimate_Npep}
\title{Get the number of theoretically observable peptides per protein}
\usage{
estimate_Npep(df, Column_Npep = NULL)
}
\arguments{
\item{df}{A data frame}

\item{Column_Npep}{column containing the number of theoretically observable peptides per protein.
If NULL try to compute the number of theoretically observable peptides using iBAQ values, 
or use molecular weight.}
}
\value{
A data frame with the column 'Npep'
}
\description{
Get the number of theoretically observable peptides per protein
}
