% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{filter_conditions.InteRactome}
\alias{filter_conditions.InteRactome}
\title{Filters conditions from an interactome}
\usage{
\method{filter_conditions}{InteRactome}(res, conditions_to_filter_out)
}
\description{
Filters conditions from an interactome
}
