% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{filter_Proteins}
\alias{filter_Proteins}
\title{Filtering of a data frame using a threshold on protein identification score and 
gene names}
\usage{
filter_Proteins(df, min_score = 0, Column_gene_name = "Gene.names")
}
\arguments{
\item{df}{A data frame}

\item{min_score}{Threshold for protein identification score}

\item{Column_gene_name}{The name of df's column containing gene names}
}
\value{
A filtered data frame
}
\description{
Filtering of a data frame using a threshold on protein identification score and 
gene names
}
