% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{geom_mean}
\alias{geom_mean}
\title{Perform the geometric mean of a numeric vector}
\usage{
geom_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{remove NA values}
}
\value{
A numeric value
}
\description{
Perform the geometric mean of a numeric vector
}
