% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCrossingScenario.R
\name{Get2DRotationMatrix}
\alias{Get2DRotationMatrix}
\title{Get a matrix M for rotating a 2D column vector X by the input parameter 
angle around the origin}
\usage{
Get2DRotationMatrix(x)
}
\arguments{
\item{x}{roating angle}
}
\value{
a 2-D matrix
}
\description{
Get a matrix M for rotating a 2D column vector X by the input parameter 
angle around the origin
}
