% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deflection.R
\name{get_deflection}
\alias{get_deflection}
\title{Compute Deflection for an Event}
\usage{
get_deflection(
  d,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)
}
\arguments{
\item{d}{inteRact-style events dataframe}

\item{equation_key}{a string corresponding to the equation key from actdata}

\item{eq_df}{if you use your own equation, this parameter should
be your equation dataframe, which (should have been reshaped by the
reshape_new_equation function prior)}

\item{dictionary_key}{a string corresponding to the dictionary from actdata you are using for cultural EPA measurements}

\item{gender}{either average, male, or female, depending on if you are using gendered equations}
}
\value{
calculated deflection for the event (single number)
}
\description{
This function calculates the deflection for ABO events contained in a dataframe
corresponding to the inteRact event data structure.
}
\examples{
d <- tibble::tibble(actor = "ceo", behavior = "advise", object = "benefactor")
d <- reshape_events_df(df = d, df_format = "wide", dictionary_key = "usfullsurveyor2015", dictionary_gender = "average")
get_deflection(d = d, equation_key = "us2010", equation_gender= "average")
}
