% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionaries.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Get Dictionary Identities}
\usage{
get_dictionary(dict_key, g)
}
\arguments{
\item{dict_key}{of dictionary, options = china1999, china2000, egypt2014, gaysex1980,
germany1989, germany2007, household1994, indiana2003, internationaldomesticrelations1981,
internet1998, japan1984, japan19892002, morocco2015, nc1978, nireland1977, ontario1980,
ontario2001, politics2003, prisonersdilemma, texas1998, uga2015, uga2015bayesactsubset,
us2010, usfullsurveyor2015, usmturk2015, usstudent2015}

\item{type}{behaviors, identities, mods}

\item{gender}{average, male, female}
}
\value{
the ACT dictionary from actdata
}
\description{
Get Dictionary Identities
}
\examples{
get_dictionary("morocco2015", "average")

}
