% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionaries.R
\name{get_equation}
\alias{get_equation}
\title{Get Equation}
\usage{
get_equation(name = NULL, type, g = NULL, eq_df = NULL)
}
\arguments{
\item{name}{name of equation. options are all equation data set keys available in the actdata package (call actdata::eqn_info() for more information)}

\item{type}{type of equation. options: emotionid, impressionabo, selfdir, traitid}

\item{gender}{gender of equation. options: f, m, av}
}
\value{
equation dataframe
}
\description{
Get Equation
}
\examples{

get_equation("nc1978", "impressionabo", "male")
}
