% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{GetInitialPositionOfCrossingRoadUser}
\alias{GetInitialPositionOfCrossingRoadUser}
\alias{GetInitialPositionOfCrossingRoadUser,Pedestrian-method}
\title{An S4 method for GetInitialPositionOfCrossingRoadUser}
\usage{
GetInitialPositionOfCrossingRoadUser(x, idx, ...)

\S4method{GetInitialPositionOfCrossingRoadUser}{Pedestrian}(x)
}
\description{
Details of the method
}
