% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{GetLongDistanceFromCrosserInitialPosToApproacherFront}
\alias{GetLongDistanceFromCrosserInitialPosToApproacherFront}
\alias{GetLongDistanceFromCrosserInitialPosToApproacherFront,Pedestrian-method}
\title{Found in GetLongDistanceFromCrosserInitialPosToApproacherFront}
\usage{
GetLongDistanceFromCrosserInitialPosToApproacherFront(x, ...)


  \S4method{GetLongDistanceFromCrosserInitialPosToApproacherFront}{Pedestrian}(x,
  SApproachingRoadUserBehaviour)
}
\arguments{
\item{x}{an object of Pedestrian or Driver}
}
\value{
a floating point scalar
}
\description{
Found in GetLongDistanceFromCrosserInitialPosToApproacherFront
}
\examples{
timeStep   <- 1/30
endTime    <- 30 ## endTime    <- 2
VTimeStamp <- seq(0, endTime, by = timeStep)
}
