% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{GetModelCrossingOnsetTimePDF}
\alias{GetModelCrossingOnsetTimePDF}
\alias{GetModelCrossingOnsetTimePDF,Pedestrian-method}
\title{Found in GetModelCrossingOnsetTimePDF.m}
\usage{
GetModelCrossingOnsetTimePDF(x, ...)

\S4method{GetModelCrossingOnsetTimePDF}{Pedestrian}(x, VTimeStamp,
  SApproachingRoadUserBehaviour)
}
\arguments{
\item{SCrossingRoadUserConstants}{a S4 class declared in model-class.R}
}
\value{
a vector perhaps
}
\description{
Linking to fGetCrossingOnsetTimePDF.
}
