% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{global_analysis.InteRactome}
\alias{global_analysis.InteRactome}
\title{Adds global variables by analysing values
across all conditions of an \code{InteRactome}}
\usage{
\method{global_analysis}{InteRactome}(res)
}
\arguments{
\item{res}{an \code{InteRactome}}
}
\value{
an \code{InteRactome} with global varaiables
}
\description{
Adds global variables by analysing values
across all conditions of an \code{InteRactome}
}
