% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{global_analysis}
\alias{global_analysis}
\title{Adds global variables by analysing values
across all conditions of an \code{InteRactome}}
\usage{
global_analysis(res)
}
\arguments{
\item{res}{an \code{InteRactome}}
}
\value{
an \code{InteRactome} with global varaiables
}
\description{
Adds global variables by analysing values
across all conditions of an \code{InteRactome}
}
