% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{identify_indirect_interactions}
\alias{identify_indirect_interactions}
\title{Identify indirect interactions by comparing stoichiometries between two interactomes.}
\usage{
identify_indirect_interactions(resA, resB, conditions = NULL)
}
\arguments{
\item{resA}{reference interactome}

\item{resB}{intermediate interactome.}

\item{conditions}{set of conditions for which stoichiometries will be compared. If \code{NULL} then all shared conditions will be used.}
}
\description{
Identify indirect interactions by comparing stoichiometries between two interactomes.
}
