% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{identify_interactors}
\alias{identify_interactors}
\title{Identify specific interactors in an \code{InteRactome}}
\usage{
identify_interactors(res, var_p_val = "p_val", p_val_thresh = 0.05,
  fold_change_thresh = 2, conditions = res$conditions,
  n_success_min = 1, consecutive_success = FALSE, p_val_breaks = c(1,
  min(1, 4 * p_val_thresh), min(1, 2 * p_val_thresh), min(1,
  p_val_thresh)), var_min_p_val = paste("min_", var_p_val, sep = ""),
  ...)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{var_p_val}{name of the p-value variable}

\item{p_val_thresh}{p-value threshold}

\item{fold_change_thresh}{fold-change threshold}

\item{conditions}{Select conditions used to identify interactors}

\item{n_success_min}{minimal number of conditions in which the interactor 
must pass the the p-value and the fold-change thresholds}

\item{consecutive_success}{logical, impose that the interactor must pass selection thresholds 
in \code{n_success_min} consecutive conditions.}

\item{...}{additionnal paramters passed to function \code{order_interactome()}}
}
\value{
an \code{InteRactome} with extra variables \code{is_interactor}, 
\code{n_success} and \code{interactor}
}
\description{
Identify specific interactors in an \code{InteRactome}
}
