% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Cbl}
\alias{Interactome_Cbl}
\title{An example Interactome computed from the 'proteinGroups_Cbl' file. 
This interactome was created using the following code :
\code{InteRact(proteinGroups_Cbl, bait_gene_name = "Cbl")}}
\format{
An InteRactome
}
\usage{
Interactome_Cbl
}
\description{
An example Interactome computed from the 'proteinGroups_Cbl' file. 
This interactome was created using the following code :
\code{InteRact(proteinGroups_Cbl, bait_gene_name = "Cbl")}
}
\keyword{datasets}
