% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Cblb}
\alias{Interactome_Cblb}
\title{Interactome of Cblb in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Cblb
}
\description{
Interactome of Cblb in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
