% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Fyb}
\alias{Interactome_Fyb}
\title{Interactome of Fyb in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Fyb
}
\description{
Interactome of Fyb in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
