% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Grb2}
\alias{Interactome_Grb2}
\title{Interactome of Grb2 in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Grb2
}
\description{
Interactome of Grb2 in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
