% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Inpp5d}
\alias{Interactome_Inpp5d}
\title{Interactome of Inpp5d in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Inpp5d
}
\description{
Interactome of Inpp5d in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
