% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Itk}
\alias{Interactome_Itk}
\title{Interactome of Itk in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Itk
}
\description{
Interactome of Itk in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
