% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Lck}
\alias{Interactome_Lck}
\title{Interactome of Lck in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Lck
}
\description{
Interactome of Lck in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
