% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Lcp2}
\alias{Interactome_Lcp2}
\title{Interactome of Lcp2 in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Lcp2
}
\description{
Interactome of Lcp2 in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
