% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Nck1}
\alias{Interactome_Nck1}
\title{Interactome of Nck1 in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Nck1
}
\description{
Interactome of Nck1 in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
