% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Nfatc2}
\alias{Interactome_Nfatc2}
\title{Interactome of Nfatc2 in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Nfatc2
}
\description{
Interactome of Nfatc2 in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
