% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Plcg1}
\alias{Interactome_Plcg1}
\title{Interactome of Plcg1 in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Plcg1
}
\description{
Interactome of Plcg1 in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
