% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Ptpn6}
\alias{Interactome_Ptpn6}
\title{Interactome of Ptpn6 in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Ptpn6
}
\description{
Interactome of Ptpn6 in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
