% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Themis}
\alias{Interactome_Themis}
\title{Interactome of Themis in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Themis
}
\description{
Interactome of Themis in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
