% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_confirm_backup.R,
%   R/maximally_confirm_behavior.R
\name{maximally_confirm_behavior}
\alias{maximally_confirm_behavior}
\title{maximally_confirm_behavior}
\usage{
maximally_confirm_behavior(
  d,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)

maximally_confirm_behavior(
  d,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)
}
\arguments{
\item{d}{events dataframe reshaped by reshape_events_dataframe function with an actor and object pairing}

\item{equation_key}{a string corresponding to the equation key from actdata}

\item{equation_gender}{either average, male, or female, depending on if you are using gendered equations}

\item{eq_df}{if you select "user supplied" for equation, this parameter should
be your equation dataframe, which (should have been reshaped by the
reshape_new_equation function prior)}

\item{actor}{lowercase string corresponding to the actor identity}

\item{beh}{lowercase string corresponding to the behavior term}

\item{object}{lowercase string corresponding to the object identity}

\item{gender}{either average, male, or female, depending on if you are using gendered equations}

\item{dictionary_key}{a string corresponding to the dictionary from actdata you are using for cultural EPA measurements}
}
\value{
3 digit EPA indicating the optimal behavior

3 digit EPA indicating the optimal behavior
}
\description{
Identify the behavior that would maximally confirm the identities of actor
and object pairing

Identify the behavior that would maximally confirm the identities of actor
and object pairing
}
\examples{
d <- tibble::tibble(actor = "ceo", object = "benefactor")
d <- reshape_events_df(df = d, df_format = "wide", dictionary_key = "usfullsurveyor2015", dictionary_gender = "average")
maximally_confirm_behavior(d = d, equation_key = "us2010", equation_gender = "average")
}
