% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{mean_analysis}
\alias{mean_analysis}
\title{Compute the mean \code{InteRactome} (on variables 'p_val', 'fold_cahnge', 'stoichio' and 'stoichio_bio')
from a list of \code{InteRactome}}
\usage{
mean_analysis(res, log = TRUE, na.rm = TRUE)
}
\arguments{
\item{res}{a list of \code{InteRactome}}

\item{log}{logical, use the geometric mean instead of the arithmetic mean}

\item{na.rm}{logical, remove NA values}
}
\description{
Compute the mean \code{InteRactome} (on variables 'p_val', 'fold_cahnge', 'stoichio' and 'stoichio_bio')
from a list of \code{InteRactome}
}
