% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{merge_conditions}
\alias{merge_conditions}
\title{Merge different conditions from different interactomes into a single dataframe}
\usage{
merge_conditions(res, selected_conditions = NULL)
}
\arguments{
\item{res}{a list of \code{InteRactome}}

\item{selected_conditions}{a character vector containing names of conditions to merge}
}
\value{
a data.frame with columns bait, names, Protein.IDs, conditions, p_val, fold_change
}
\description{
Merge different conditions from different interactomes into a single dataframe
}
