% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{merge_duplicate_groups}
\alias{merge_duplicate_groups}
\title{Merge protein groups with the same gene name.}
\usage{
merge_duplicate_groups(df, idx_col = NULL, merge_column = "gene_name",
  sum_intensities = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{idx_col}{idx of columns for which values will be merged across protein groups}

\item{merge_column}{column to identify rows to be be merged}

\item{sum_intensities}{Logical. Sum intensities across merged groups.}
}
\value{
A merged data frame
}
\description{
Merge protein groups with the same gene name.
}
