% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_identity.R
\name{modify_identity}
\alias{modify_identity}
\title{Modify Identity}
\usage{
modify_identity(id_info, eq_info)
}
\arguments{
\item{id_info}{a dataframe from reshape events that should be only
actor_modifier and actor elements or object_modifier and object elements}

\item{eq_info}{the equation information you are using, should be in the
form {name}_{gender}}
}
\value{
three digit EPA profile of the modified identity
}
\description{
A function that applies the modifier equations to find the EPA location for modifier
\itemize{
\item identity
}
}
\examples{
mi <- tibble::tibble(actor_modifier = c("adventurous", "adventurous"),actor = c("man", "woman"))

mi_reshaped <- reshape_events_df(df = mi, df_format = "wide",
dictionary_key = "indiana2003", dictionary_gender = "male")
mod_id <- mi_reshaped \%>\%
dplyr::group_by(event_id) \%>\% tidyr::nest() \%>\%
dplyr::mutate(eq_info = "nc1978_male") \%>\%
dplyr::mutate(mod_id = purrr::map2(data, eq_info, modify_identity))

}
