% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{moving_average}
\alias{moving_average}
\title{Performs a running average on a numeric vector}
\usage{
moving_average(x, n)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{integer, radius of the moving avergae (number of points extending on each side of the 
center point on which the average is computed)}
}
\value{
a smoothed numeric vector
}
\description{
Performs a running average on a numeric vector
}
