% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{multiplot}
\alias{multiplot}
\title{Arrange multiple plots in a given layout}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{A list of plots}

\item{cols}{Number of columns to use (if \code{layout} is NULL)}

\item{layout}{layout of plots (a matrix representing the position of the different plots).
For instance layout = matrix(c(1,1,2,3), ncol=2, nrow=2) will plot the first plot in the upper 
half while the bottom half will be equally split between plots 2 and 3.}
}
\value{
A plot
}
\description{
Arrange multiple plots in a given layout
}
