% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{normalize_interactome}
\alias{normalize_interactome}
\title{Normalize the log fold change by its standard deviation for each condition}
\usage{
normalize_interactome(res)
}
\arguments{
\item{res}{an \code{InteRactome}}
}
\value{
an \code{InteRactome} with the additional variable \code{norm_log_fold_change}
}
\description{
Normalize the log fold change by its standard deviation for each condition
}
