% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_behavior.R
\name{optimal_behavior}
\alias{optimal_behavior}
\title{Calculate the Optimal Behavior for the Actor following an Event}
\usage{
optimal_behavior(
  d,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)
}
\arguments{
\item{d}{data that has been reshaped by the reshape_events_df function}

\item{equation_key}{a string corresponding to the equation key from actdata}

\item{equation_gender}{either average, male, or female, depending on if you are using gendered equations}
}
\value{
3 digit EPA indicating the optimal behavior
}
\description{
Calculate the Optimal Behavior for the Actor following an Event
}
\examples{

opt_behavior_example <- tibble::tibble(actor = "teenager", behavior = "beam_at",object = "friend")
opt_behavior_df <- reshape_events_df(df = opt_behavior_example, df_format = "wide", dictionary_key = "indiana2003", dictionary_gender = "male")
opt_b <- optimal_behavior(d = opt_behavior_df, equation_key = "nc1978", equation_gender = "male")



}
