% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{Pedestrian-class}
\alias{Pedestrian-class}
\title{An S4 class for a Pedestrian}
\description{
Details of the class...
}
\section{Slots}{

\describe{
\item{\code{sCountry}}{a character string to represent which country.}

\item{\code{sScenarioName}}{a character string.}

\item{\code{fGetCrossingTrajectory}}{a function.}

\item{\code{clength}}{a floating point scalar for crossing.}

\item{\code{cwidth}}{a floating point scalar for crossing}

\item{\code{intersectionLaneWidth}}{a floating point scalar.}

\item{\code{SModelSpecificConstants}}{a list.}

\item{\code{alength}}{a floating point scalar for approaching.}

\item{\code{awidth}}{a floating point scalar for approaching.}

\item{\code{fGetApproachBehaviour}}{a function}

\item{\code{fAdaptToCrossingBehaviour}}{a function}

\item{\code{initialSpeed}}{a floating point scalar.}

\item{\code{initialTTA}}{a floating point scalar}

\item{\code{decelerationOnsetTTA}}{a floating point scalar}

\item{\code{stopAtDistanceFromConflict}}{a floating point scalar}

\item{\code{eHMIOnsetTTA}}{a NaN.}

\item{\code{desiredPostEncroachmentTime}}{a floating point scalar}

\item{\code{maxAcceleration}}{a floating point scalar}
}}

