% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_annotation_results}
\alias{plot_annotation_results}
\title{Plot the result of the annotation enrichment analysis}
\usage{
plot_annotation_results(df, var_p_val = "fdr",
  fold_change_max_plot = 4, save_file = NULL, ...)
}
\arguments{
\item{df}{a formatted data.frame obtained by the function \code{annotation_enrichment_analysis()}}

\item{p_val_max}{threshold for the enrichment p-value}

\item{method_adjust_p_val}{method to adjust p-value for multiple comparisons}

\item{fold_change_min}{threshold for the enrichment fold-change}

\item{N_annot_min}{minimum number of elements that are annotated in the foreground set}
}
\value{
a plot
}
\description{
Plot the result of the annotation enrichment analysis
}
