% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Plot protein intensities per biological replicate and background}
\usage{
plot_comparison(res, name, conditions, textsize = 3, test = "t.test",
  test.args = list(paired = FALSE), map_signif_level = c(`***` = 0.001, `**`
  = 0.01, `*` = 0.05), position = "position_jitter",
  position.args = list(width = 0.3, height = 0))
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{name}{name of the protein to display}

\item{conditions}{set of conditions to display}

\item{textsize}{size of labels corresponding to significance levels}

\item{test}{name of the test function to compare intensities between background}

\item{test.args}{arguments passed to function \code{test}}

\item{map_signif_level}{named vector with labels and corresponding significance levels}

\item{position}{name of the function used to position data points}

\item{position.args}{arguments passed to function \code{position}}
}
\value{
a plot
}
\description{
Plot protein intensities per biological replicate and background
}
