% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{plot_correlation_network}
\alias{plot_correlation_network}
\title{Plot an interactive correlation network with communities highlighted}
\usage{
plot_correlation_network(res, idx = NULL, df_corr = NULL,
  r_corr_thresh = 0.8, p_val_thresh = 0.05)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{idx}{indexes of the set of proteins in \code{res} for which correlations will be computed}

\item{df_corr}{a data.frame with columns 'r_corr' and 'p_corr'. Has priority over parameters \code{res} and \code{idx}.}

\item{r_corr_thresh}{threshold for variable 'r_corr' (min)}

\item{p_val_thresh}{threshold for ariable 'p_corr' (max)}
}
\value{
an interactive networkD3 plot
}
\description{
Plot an interactive correlation network with communities highlighted
}
