% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{plot_QC}
\alias{plot_QC}
\title{Quality check plots for preprocessed AP-MS data}
\usage{
plot_QC(prep_data)
}
\arguments{
\item{prep_data}{preprocessed data as obtained using function \code{preprocess_data()}}
}
\value{
Several QC plots
}
\description{
Quality check plots for preprocessed AP-MS data
}
