% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{plot_stoichio}
\alias{plot_stoichio}
\title{Plot interaction stoichiometries per biological replicate}
\usage{
plot_stoichio(res, name, conditions = res$conditions,
  ref_condition = conditions[1], test = "t.test",
  test.args = list(paired = TRUE), map_signif_level = c(`***` = 0.001,
  `**` = 0.01, `*` = 0.05), save_file = NULL)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{name}{name of the protein to display}

\item{conditions}{set of conditions to display}

\item{ref_condition}{name of the reference condition for all \code{test}}

\item{test}{name of the test function to compare sttoichiometries between conditions}

\item{test.args}{arguments passed to function \code{test}}

\item{map_signif_level}{named vector with labels and corresponding significance levels}

\item{save_file}{path of output file (.pdf)}
}
\value{
a plot
}
\description{
Plot interaction stoichiometries per biological replicate
}
