% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_volcanos}
\alias{plot_volcanos}
\title{Plot protein enrichement fold-change versus p-value}
\usage{
plot_volcanos(
  res = NULL,
  data = NULL,
  names = NULL,
  labels = NULL,
  N_print = 15,
  conditions = NULL,
  p_val_thresh = 0.05,
  fold_change_thresh = 2,
  x0 = NULL,
  c = NULL,
  save_file = NULL,
  xlim = NULL,
  ylim = NULL,
  asinh_transform = TRUE,
  norm = FALSE,
  both_sides = FALSE,
  theme_name = "theme_gray"
)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{data}{data.frame with columns 'names', 'p_val' and 'fold_change'}

\item{names}{Names of proteins highlighted}

\item{labels}{labels for proteins in plot. Must the same length as \code{res$names}}

\item{N_print}{maximum of protein labels to display}

\item{conditions}{conditions to plot}

\item{p_val_thresh}{threshold on p-value to display}

\item{fold_change_thresh}{threshold on fold-change to display}

\item{x0}{parameters x0 of the line dividing the volcano plot according to \code{f(x) = c / (|x|-x0)}.
Ignored unless parameters \code{p_val_thresh} and \code{fold_change_thresh} are set to \code{NULL}}

\item{c}{parameters c of the line dividing the volcano plot according to \code{f(x) = c / (|x|-x0)}.
Ignored unless parameters \code{p_val_thresh} and \code{fold_change_thresh} are set to \code{NULL}}

\item{save_file}{path of output file (.pdf)}

\item{xlim}{range of x values}

\item{ylim}{range of y values}

\item{asinh_transform}{logical, display asinh(log10(p-value)) on the y-axis}

\item{norm}{Use normalized fold-changes}

\item{both_sides}{logical. Shading on right and left upper graphs.}

\item{theme_name}{name of the ggplot2 theme function to use ('theme_gray' by default)}
}
\value{
a plot
}
\description{
Plot protein enrichement fold-change versus p-value
}
