% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocessing of raw data}
\usage{
preprocess_data(df, Column_gene_name = "Gene.names", Column_score = "Score",
  Column_ID = "Protein.IDs", Column_Npep = NULL,
  Column_intensity_pattern = "^Intensity.", bait_gene_name,
  bckg_bait = bait_gene_name, bckg_ctrl = "WT", log = TRUE,
  condition = NULL, filter_time = NULL, filter_bio = NULL,
  filter_tech = NULL, ...)
}
\arguments{
\item{Column_gene_name}{Column with gene names}

\item{Column_score}{Column with protein identification score}

\item{Column_ID}{Column with protein IDs}

\item{Column_Npep}{Column with number of theoretically observable peptides per protein}

\item{Column_intensity_pattern}{Pattern (regular exrpression) used to identfy df's columns containing protein intensity values}

\item{bait_gene_name}{The gene name of the bait}

\item{bckg_bait}{Τhe name of the bait background as found in \code{condition$bckg} (see below)}

\item{bckg_ctrl}{Τhe name of the control background as found in \code{condition$bckg} (see below)}

\item{log}{logical, use geometric mean to average technical replicates}

\item{condition}{data.frame with columns "sample", bckg", "bio", "time" and "tech" indicating 
for each intensity column ("sample") its corresponding background ("bckg"), biologicla replicate ("bio), 
experimental condition ("tine) and technical replicate ("tech).}

\item{filter_time}{vector of experimental conditions to exclude from analysis}

\item{filter_bio}{vector of biological replicates to exclude from analysis}

\item{filter_tech}{vector of technical replicates to exclude from analysis}

\item{...}{Additional parameters passed to function \code{identify_conditions}}
}
\description{
Preprocessing of raw data
}
