% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteome_CD4}
\alias{proteome_CD4}
\title{Proteome of antigen-experienced conventional CD4+ T cells.}
\format{A data frame with 6355 rows (one per protein ID) and 23 variables.}
\usage{
proteome_CD4
}
\description{
The cellular abundance of proteins extracted from lysates 
of CD4+ T cells lysates are shown as numbers of copies per cell
}
\keyword{datasets}
