% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteome_CD4_expanded}
\alias{proteome_CD4_expanded}
\title{Proteome of CD4+ T cells from Cas9-GFP mice after expansion and transfection with a CRISPR gRNA.}
\format{A data frame with 19690 rows (one per protein ID) and 20 variables.}
\usage{
proteome_CD4_expanded
}
\description{
The cellular abundance of proteins extracted from lysates 
of CD4+ T cells lysates are shown as numbers of copies per cell
}
\keyword{datasets}
