% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteome_Jurkat}
\alias{proteome_Jurkat}
\title{Proteome of Jurkat T cells}
\format{A data frame with 4399 rows (one per protein ID) and 20 variables.}
\usage{
proteome_Jurkat
}
\description{
The cellular abundance of proteins extracted from lysates 
of Jurkat T cells lysates are shown as numbers of copies per cell
}
\keyword{datasets}
