% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify_actor.R
\name{reidentify_actor}
\alias{reidentify_actor}
\title{Function to Generate the Re-identification of the Actor}
\usage{
reidentify_actor(
  data,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)
}
\arguments{
\item{df}{data that has been reshaped by the events_df}

\item{equation_info}{is a string that corresponds to "{equationkey}_{gender}" from actdata}
}
\value{
dataframe with 3 columns corresponding to the EPA of optimal actor identity relabel
}
\description{
Function to Generate the Re-identification of the Actor
}
\examples{
d <- tibble::tibble(actor = "ceo", behavior = "advise", object = "benefactor")
d <- reshape_events_df(df = d, df_format = "wide", dictionary_key = "usfullsurveyor2015", dictionary_gender = "average")
reidentify_actor(data = d, equation_key = "us2010", equation_gender = "average")
}
