% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{rescale_median}
\alias{rescale_median}
\title{Normalize data frame by columns using the median}
\usage{
rescale_median(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A normalized data frame
}
\description{
Normalize data frame by columns using the median
}
