% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_events_df.R
\name{reshape_events_df}
\alias{reshape_events_df}
\title{Reshape Events DF}
\usage{
reshape_events_df(
  df,
  df_format = c("wide", "long"),
  id_column = NULL,
  dictionary_key,
  dictionary_gender = c("male", "female", "average")
)
}
\arguments{
\item{df}{dataframe of events}

\item{df_format}{whether your events dataframe is wide (a column for actor, behavior, and object)
or long (multiple rows for each event)}

\item{id_column}{if your df is in long format, you must supply the column that identifies the individual event}

\item{dictionary_key}{the actdata dictionary key you are using for your analysis}

\item{dictionary_gender}{the gender EPA ratings you are using for your analysis. Should be male, female, or average}
}
\value{

}
\description{
Reshape Events DF
}
\examples{

d <- tibble::tibble(actor = "ceo", behavior = "advise", object = "benefactor")
reshape_events_df(df = d, df_format = "wide", dictionary_key = "usfullsurveyor2015", dictionary_gender = "average")
}
