% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_equations_functions.R
\name{reshape_new_equation}
\alias{reshape_new_equation}
\title{Reshape New Equation}
\usage{
reshape_new_equation(eq_df)
}
\arguments{
\item{eq_df}{a data frame containing the user-supplied equation. Should follow
the same structure as those found in Interact.jar (10 columns, the first of which
should have the Z plus 9 1s or 0s indicating the elements of the corresponding
coefficient structure).}
}
\value{
eq_coef_info dataframe that is in the correct form for use in other functions
}
\description{
Reshape New Equation
}
