% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{restrict_network_degree}
\alias{restrict_network_degree}
\title{Limit the number of edges per node within a network}
\usage{
restrict_network_degree(df_corr, source = "name_1", target = "name_2",
  nmax = NULL, strict = TRUE)
}
\arguments{
\item{df_corr}{a data.frame}

\item{source}{column with source nodes}

\item{target}{column with target nodes}

\item{nmax}{integer, limits the number of edges per node}

\item{strict}{logical, if TRUE, ensures a strict upper bound on the maximal degree}
}
\value{
a network with a maximum degree lower than \code{nmax}
}
\description{
Limit the number of edges per node within a network
}
