% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{row_mean}
\alias{row_mean}
\title{Compute the mean by row}
\usage{
row_mean(df, na.rm = TRUE, log = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{log}{option to use geometric mean}
}
\value{
A numeric vector
}
\description{
Compute the mean by row
}
