% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{row_sd}
\alias{row_sd}
\title{Compute the standard deviation by rows}
\usage{
row_sd(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
A numeric value
}
\description{
Compute the standard deviation by rows
}
