% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{row_stoichio}
\alias{row_stoichio}
\title{Compute the stoichiometry of interaction using the method described in ...}
\usage{
row_stoichio(df, idx_group_1, idx_group_2, idx_bait, Npep, log = TRUE,
  substract_ctrl = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{idx_group_1}{column indexes corresponding to the first group (bait background)}

\item{idx_group_2}{column indexes corresponding to the second group (ctrl background)}

\item{idx_bait}{row index for the bait protein}

\item{Npep}{numeric vector containing the number of theoretically observable peptides for each protein}

\item{log}{option to use the geometric mean instead of the arithmetic mean}
}
\value{
A numeric vector of interaction stoichiometries
}
\description{
Compute the stoichiometry of interaction using the method described in ...
}
