% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCrossingScenario.R
\name{SimulateCrossingScenario}
\alias{SimulateCrossingScenario}
\title{SimulateCrossingScenario - run interACT road crossing model simulation}
\usage{
SimulateCrossingScenario(x, VTimeStamp)
}
\arguments{
\item{x}{a Pedestrian or a Driver}

\item{VTimeStamp}{a time vector}
}
\description{
SimulateCrossingScenario - run interACT road crossing model simulation
}
