% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{smooth}
\alias{smooth}
\title{Smooth, using a moving average across conditions, selected variables of an \code{InteRactome}}
\usage{
smooth(x, ...)
}
\description{
Smooth, using a moving average across conditions, selected variables of an \code{InteRactome}
}
