% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{summary_protein}
\alias{summary_protein}
\title{Create a summary for selected proteins in an \code{InteRactome}}
\usage{
summary_protein(res, name, idx = NULL)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{name}{protein name}

\item{idx}{indexes of the proteins to display. Overrides \code{name}}
}
\value{
a data.frame
}
\description{
Create a summary for selected proteins in an \code{InteRactome}
}
