% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{summary_table}
\alias{summary_table}
\title{Create a summary table for an \code{InteRactome}}
\usage{
summary_table(res, add_columns = names(res))
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{add_columns}{names of the variables to display in the summary table}
}
\value{
a data.frame
}
\description{
Create a summary table for an \code{InteRactome}
}
