% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_distr.R
\name{threshold_distr}
\alias{threshold_distr}
\title{Implementation of the threshold distribution model (TDM) described in
interACT D2.2}
\usage{
threshold_distr(ts, distance, speed, ehmi, p)
}
\arguments{
\item{ts}{a time vector with a fixed interval}

\item{distance}{a double vector storing distance}

\item{speed}{a double vector storing speed}

\item{ehmi}{a logical vector}

\item{p}{a list storing parameters}
}
\description{
Implementation of the threshold distribution model (TDM) described in
interACT D2.2
}
