% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_1978}
\alias{us_1978}
\title{US 1978 Male Equations}
\format{
A data frame with 20 rows and 22 variables:
\describe{
\item{coef_name}{indicator of what elements the beta is related to}
\item{postAE}{coefficients to estimate the transient impression of the Actor's Evaluation}
\item{postAP}{point coefficients to estimate the transient impression of the Actor's Potency}
\item{postAA}{point coefficients to estimate the transient impression of the Actor's Activity}
\item{postBE}{point coefficients to estimate the transient impression of the Behavior's Evaluation}
\item{postBP}{point coefficients to estimate the transient impression of the Behavior's Potency}
\item{postBA}{point coefficients to estimate the transient impression of the Behavior's Activity}
\item{postOE}{point coefficients to estimate the transient impression of the Object's Evaluation}
\item{postOP}{point coefficients to estimate the transient impression of the Object's Potency}
\item{postOA}{point coefficients to estimate the transient impression of the Object's Activity}
\item{A}{splitting apart Z into element A}
\item{B}{splitting apart Z into element B}
\item{O}{splitting apart Z into element O}
\item{AE}{1 or 0 indicator whether AE is involved in estimation}
\item{AP}{1 or 0 indicator whether AP is involved in estimation}
\item{AA}{1 or 0 indicator whether AA is involved in estimation}
\item{BE}{1 or 0 indicator whether BE is involved in estimation}
\item{BP}{1 or 0 indicator whether BP is involved in estimation}
\item{BA}{1 or 0 indicator whether BA is involved in estimation}
\item{OE}{1 or 0 indicator whether OE is involved in estimation}
\item{OP}{1 or 0 indicator whether OP is involved in estimation}
\item{OA}{1 or 0 indicator whether OA is involved in estimation}
}
}
\source{
\url{http://www.tandfonline.com/doi/pdf/10.1080/0022250X.1987.9990026}
}
\usage{
us_1978
}
\description{
A dataframe of the coefficients for the impression formation equations that was estimated in 1978.
A new version of the equations are currently being estimated. When this has been completed, the new
dataframe will be added to this package.
}
\details{
Please cite as:<FIGURE THIS OUT>
}
\keyword{datasets}
