% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{average_technical_replicates}
\alias{average_technical_replicates}
\title{Average protein intensities over technical replicates}
\usage{
average_technical_replicates(df, cond, log = TRUE)
}
\arguments{
\item{df}{A data frame of protein intensities}

\item{cond}{A data frame containing the description of df's columns (i.e "bckg", "time", "bio"  and "tech") 
as returned by function \code{identify_conditions()}}

\item{log}{use geometric mean}
}
\value{
A list containing :

\code{Intensity}, a data frame of protein intensities averaged over technical replicates;

\code{conditions}, a data frame containing the description of \code{Intensity}'s columns
}
\description{
Average protein intensities over technical replicates
}
\examples{
#load data :
data("proteinGroups_Cbl")
df <- proteinGroups_Cbl
cond <- identify_conditions(df)
Column_intensity_pattern <- "^Intensity."
df_int <- df[ , grep(Column_intensity_pattern, colnames(df))]
}
