% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_deflection.R
\name{batch_deflection}
\alias{batch_deflection}
\title{Batch Deflection}
\usage{
batch_deflection(df)
}
\arguments{
\item{df}{a dataframe that must have three columns named: actor, behavior, and object
that correspond to the elements of your event. These must match with identities and behaviors
in the US 2015 dictionary.}
}
\value{
a dataframe with a column indicating the total deflection for the event
}
\description{
Batch Deflection
}
