% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_element_deflection.R
\name{batch_element_deflection}
\alias{batch_element_deflection}
\title{Batch Element Deflection}
\usage{
batch_element_deflection(df)
}
\arguments{
\item{df}{a dataframe that must have three columns named: actor, behavior, and object
that correspond to the elements of your event. These must match with identities and behaviors
in the US 2015 dictionary.}
}
\value{
a dataframe in long format with a variable indicating the amount of deflection produced by each
element
}
\description{
Batch Element Deflection
}
