% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_emotion.R
\name{characteristic_emotion}
\alias{characteristic_emotion}
\title{Characteristic Emotion}
\usage{
characteristic_emotion(id_info, equation_info)
}
\arguments{
\item{id_info}{a df from reshape_events_df with EPA information of the identity}

\item{equation_info}{is a string that corresponds to "{equationkey}_{gender}" from actdata}
}
\value{
df with a 3 digit EPA profile of characteristic emotion for the identity
}
\description{
Characteristic Emotion
}
\examples{
id <- tibble::tibble(actor = "brute")
id_info <- reshape_events_df(id, df_format = "wide", dictionary_key = "usfullsurveyor2015", dictionary_gender = "average")
characteristic_emotion(id_info, equation_info = "us2010_male")
}
