% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristic_emotion.R
\name{characteristic_emotion}
\alias{characteristic_emotion}
\title{Characteristic Emotion}
\usage{
characteristic_emotion(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Characteristic Emotion
}
