% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{compare_stoichio}
\alias{compare_stoichio}
\title{Compare stoichiometries between two conditions using a ttest}
\usage{
compare_stoichio(res, names = res$names,
  ref_condition = res$conditions[1],
  test_conditions = setdiff(res$conditions, ref_condition),
  p_val_thresh = 0.05, fold_change_thresh = 3, ...)
}
\description{
Compare stoichiometries between two conditions using a ttest
}
