% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{compute_correlations}
\alias{compute_correlations}
\title{Compute correlation in protein recruitment}
\usage{
compute_correlations(res, idx = NULL, log = FALSE, nmax = NULL)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{idx}{indexes of the set of proteins for which correlations will be computed}

\item{log}{logical, use log-transformed stoichiometries}

\item{nmax}{integer, limits the number of connections per node}
}
\value{
a data.frame with protein correlation information 
(correlation coefficient in column 'r_corr' and associated p-value in column 'p-corr')
}
\description{
Compute correlation in protein recruitment from interaction stoichiometries 
computed across all conditions for each biological replicate. Pearson correlations are used.
}
