% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_h_matrix.R
\name{construct_h_matrix}
\alias{construct_h_matrix}
\title{Construct H Matrix}
\usage{
construct_h_matrix(eq)
}
\arguments{
\item{equation}{which equation to use - you can either set it to "us" for the
us 1978 equations, or "user supplied")}

\item{eq_df}{if you select "user supplied" for equation, this parameter should
be your equation dataframe, which (should have been reshaped by the
reshape_new_equation function prior)}
}
\description{
Construct H Matrix
}
