% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_select_mat.R
\name{create_select_mat}
\alias{create_select_mat}
\title{Convenience Function for Creating Selection Matrix Depending on the Term}
\usage{
create_select_mat(term, gender, equation_key, eq_df = NULL)
}
\arguments{
\item{term, }{string indicating whether you are selection for actor, behavior, or object}

\item{equation_key}{}
}
\value{

}
\description{
Convenience Function for Creating Selection Matrix Depending on the Term
}
