% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{dot_plot}
\alias{dot_plot}
\title{Dot plot representation of matrices}
\usage{
dot_plot(Dot_Size, Dot_Color = NULL, title = "Dot Plot",
  size_range = c(0, 5.5), size_limits = range(Dot_Size),
  size_breaks = NULL, size_var = "size", color_var = "color",
  color_values = c("red", "purple", "blue", "black"),
  size_label_y = NULL, size_label_x = NULL)
}
\arguments{
\item{Dot_Size}{a matrix of dot sizes}

\item{Dot_Color}{a matrix of dot colors (optionnal)}

\item{title}{plot title}

\item{size_range}{range of dot sizes to display}

\item{size_var}{name of the variable corresponding to dot size}

\item{color_var}{name of the variable corresponding to dot color}

\item{color_values}{values parameter passed to \code{scale_color_manual()}}

\item{size_label_y}{size of y-axis labels}

\item{size_label_x}{size of x-axis labels}
}
\value{
a plot
}
\description{
Dot plot representation of matrices
}
