% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_terms.R
\name{extract_terms}
\alias{extract_terms}
\title{Create the I matrix from Heise (2010) that corresponds to ABO elements that are not the one being solved for in optimal functions}
\usage{
extract_terms(elem, equation_key, gender, t_imp)
}
\arguments{
\item{elem}{string either A, B, or O}

\item{t_imp}{the dataframe containing the result of the transient impression from the event}

\item{equation}{key}
}
\value{
29 x 29 matrix
}
\description{
Create the I matrix from Heise (2010) that corresponds to ABO elements that are not the one being solved for in optimal functions
}
