% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{filter_Proteins}
\alias{filter_Proteins}
\title{Filtering of a data frame using a threshold on protein identification score and 
gene names}
\usage{
filter_Proteins(
  df,
  min_score = 0,
  filter_gene_name = TRUE,
  Column_ID = "Protein.IDs",
  Column_gene_name = "Gene.names",
  Column_score = "Score",
  split_param = ";"
)
}
\arguments{
\item{df}{A data frame}

\item{min_score}{Threshold for protein identification score}

\item{filter_gene_name}{logical, filter out proteins withy empty gene name}

\item{Column_ID}{Column with protein IDs}

\item{Column_gene_name}{The name of df's column containing gene names}

\item{Column_score}{The name of df's column containing protein identification score}

\item{split_param}{Character used to split gene names into substrings.}
}
\value{
A filtered data frame. Contains an extra column with the first substring of the column \code{Column_gene_name}
}
\description{
Filtering of a data frame using a threshold on protein identification score and 
gene names
}
