% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deflection.R
\name{get_deflection}
\alias{get_deflection}
\title{Compute Deflection for an Event}
\usage{
get_deflection(
  act,
  beh,
  obj,
  dictionary = "us",
  equation = c("us", "user supplied"),
  eq_df = NULL
)
}
\arguments{
\item{act}{lowercase string corresponding to the actor identity}

\item{beh}{lowercase string corresponding to the behavior term}

\item{obj}{lowercase string corresponding to the object identity}

\item{dictionary}{which dictionary to use, currently set to "us"}

\item{equation}{which equation to use - you can either set it to "us" for the
us 1978 equations, or "user supplied")}

\item{eq_df}{if you select "user supplied" for equation, this parameter should
be your equation dataframe, which (should have been reshaped by the
reshape_new_equation function prior)}
}
\value{
calculated deflection for the event (single number)
}
\description{
This function calculates the deflection for an Actor, Behavior, Object event.
It assumes that the first input is an identity corresponding to the actor,
the second, the behavior, and last, the object. Each of these terms must be
in the US 2015 dictionary.
}
\examples{

get_deflection("ceo", "advise", "benefactor", equation = "us")
}
