% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionaries.R
\name{get_equation}
\alias{get_equation}
\title{Get Equation}
\usage{
get_equation(name, type, gender)
}
\arguments{
\item{name}{name of equation. options: canada1985, canada20012003, china2000,
egypt2014, germany2007, japan1984, morocco2015, nc1978, us2010}

\item{type}{type of equation. options: emotionid, impressionabo, selfdir, traitid}

\item{gender}{gender of equation. options: f, m, av}
}
\value{

}
\description{
Get Equation
}
