% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{GetModelCrossingOnsetTimePDF}
\alias{GetModelCrossingOnsetTimePDF}
\alias{GetModelCrossingOnsetTimePDF,Pedestrian-method}
\title{Found in GetModelCrossingOnsetTimePDF.m}
\usage{
GetModelCrossingOnsetTimePDF(x, ...)

\S4method{GetModelCrossingOnsetTimePDF}{Pedestrian}(x, VTimeStamp,
  SApproachingRoadUserBehaviour)
}
\arguments{
\item{x}{a Pedestrian or a Driver object}

\item{SApproachingRoadUserBehaviour}{a Behaviour object storing approaching
road user behaviour.}

\item{VTimeStampe}{a time vecotr}
}
\value{
a vector perhaps?
}
\description{
Linking to fGetCrossingOnsetTimePDF.
}
