% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{identify_conditions}
\alias{identify_conditions}
\title{Identify conditions (background, time of stimulation, biological and technical replicates) 
from column names}
\usage{
identify_conditions(df, Column_intensity_pattern = "^Intensity.",
  split = "_", bckg_pos = 1, bio_pos = 3, time_pos = 2,
  tech_pos = 4)
}
\arguments{
\item{df}{A dataframe containing protein intensities. By default, protein intensity column names start by "Intensity." 
(use parameter \code{Column_intensity_pattern} to change)}

\item{Column_intensity_pattern}{Pattern (regular exrpression) used to identfy df's columns containing protein intensity values}

\item{split}{Character used to split column names into substrings}

\item{bckg_pos}{Position of the sample background in splitted column names}

\item{bio_pos}{Position of the sample biological replicate in splitted column names}

\item{time_pos}{Position of the sample experimental condition in splitted column names}

\item{tech_pos}{Position of the sample technical replicate in splitted column names}
}
\value{
a data frame describing experimental samples in terms of background, 
biological and technical replicates, and experimental conditions
}
\description{
Identify conditions (background, time of stimulation, biological and technical replicates) 
from column names
}
\examples{
#load data :
data("proteinGroups_Cbl")
# You can identify columns and their description separately using `identify_conditions()`
cond <- identify_conditions(proteinGroups_Cbl)
print.data.frame(cond)
# and use it as parameters for function InteRact()
res <- InteRact(proteinGroups_Cbl, bait_gene_name = "Cbl", condition = cond)
}
