% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{identify_conditions}
\alias{identify_conditions}
\title{Identify conditions (background, time of stimulation, biological and technical replicates) 
from column names}
\usage{
identify_conditions(df, Column_intensity_pattern = "^Intensity.", bckg_bait,
  bckg_ctrl = "WT", preffix_bio = "S", preffix_tech = "R",
  preffix_time = "")
}
\description{
Identify conditions (background, time of stimulation, biological and technical replicates) 
from column names
}
