% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{InteRact}
\alias{InteRact}
\title{Analysis of AP-MS data}
\usage{
InteRact(df, updateProgress = NULL, N_rep = 3, quantile_rep = 0.05,
  pool_background = TRUE, log_test = TRUE, log_stoichio = TRUE,
  log_mean = TRUE, by_conditions = TRUE, substract_ctrl = TRUE,
  preprocess_df = NULL, ...)
}
\arguments{
\item{df}{A dataframe containing protein intensities. By default, protein intensity column names start by "Intensity." 
(use parameter \code{Column_intensity_pattern} to change)}

\item{updateProgress}{function to show progress bar in shiny app}

\item{N_rep}{Number of iterations for the replacement of missing values}

\item{quantile_rep}{Numeric value between 0 and 1. Quantile of the distribution of mean intensities 
in the control background used to replace missing values.}

\item{pool_background}{option to use all control background conditions as one control group for all conditions}

\item{log_test}{logical, perform t-test on log transform intensities}

\item{log_stoichio}{logical, use the geometric mean instead of the arithmetic mean to compute stoichiometries}

\item{log_mean}{logical, use the geometric mean instead of the arithmetic mean to compute the mean \code{InteRactome}}

\item{by_conditions}{option to perform the comparison between bait and control group for each condition}

\item{substract_ctrl}{logical, substract ctrl intensities in the calculation of stoichiometries}

\item{preprocess_df}{list obtained by  the function \code{preprocess_data()}}

\item{...}{Additional parameters passed to function \code{preprocess_data()} and \code{identify_conditions}}
}
\value{
a list containing the preprocessed data and on object of class \code{InteRactome}, i.e a list including the following elements :

\code{conditions} : a vector of experimental conditions.

\code{names} : a vector of names (by default gene names are used).

\code{p_val} : a list of vectors containing the p values associated to each experimental condition.

\code{fold_change} : a list of vectors containing the fold change associated to each experimental condition.

\code{...} : other variables.
}
\description{
Analysis of AP-MS data
}
\examples{
#load data :
data("proteinGroups_Cbl")
#Run InteRact with default parameters
res <- InteRact(proteinGroups_Cbl, bait_gene_name = "Cbl")$Interactome

#You now have an \\code{InteRactome}. See its elements.
class(res)
names(res)
#Generate volcano plots
plot_volcanos(res)

#Identify specific interactors
res <- identify_interactors(res, p_val_thresh = 0.05, fold_change_thresh = 2)

#Visualize interaction kinetics
plot_per_condition(res)

# Append protein abundance information
res <- merge_proteome(res)
# Append annotations
annot <- get_annotations(res)
res <- append_annotations(res,  annot)
#Create a summary data frame
sum_tbl <- summary_table(res)
}
\author{
Guillaume Voisinne
}
