% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Interactome_Cbl}
\alias{Interactome_Cbl}
\title{Interactome of Cbl in effector CD4+ T cells in the course of TCR stimulation}
\format{An \code{InteRactome}}
\usage{
Interactome_Cbl
}
\description{
Interactome of Cbl in effector CD4+ T cells in the course of TCR stimulation
}
\keyword{datasets}
