% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_confirm_backup.R,
%   R/maximally_confirm_behavior.R
\name{maximally_confirm_behavior}
\alias{maximally_confirm_behavior}
\title{maximally_confirm_behavior}
\usage{
maximally_confirm_behavior(df, equation_info)

maximally_confirm_behavior(df, equation_info)
}
\arguments{
\item{actor}{lowercase string corresponding to the actor identity}

\item{beh}{lowercase string corresponding to the behavior term}

\item{object}{lowercase string corresponding to the object identity}

\item{dictionary_key}{a string corresponding to the dictionary from actdata you are using for cultural EPA measurements}

\item{gender}{either average, male, or female, depending on if you are using gendered equations}

\item{equation_key}{a string corresponding to the equation key from actdata}

\item{eq_df}{if you select "user supplied" for equation, this parameter should
be your equation dataframe, which (should have been reshaped by the
reshape_new_equation function prior)}
}
\value{
3 digit EPA indicating the optimal behavior

3 digit EPA indicating the optimal behavior
}
\description{
Identify the behavior that would maximally confirm the identities of actor
and object pairing

Identify the behavior that would maximally confirm the identities of actor
and object pairing
}
