% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximally_confirm_behavior.R
\name{maximally_confirm_behavior}
\alias{maximally_confirm_behavior}
\title{maximally_confirm_behavior}
\usage{
maximally_confirm_behavior(actor, object, dictionary, equation, eq_df = NULL)
}
\arguments{
\item{actor}{lowercase string corresponding to the actor identity}

\item{object}{lowercase string corresponding to the object identity}

\item{dictionary}{which dictionary to use, currently set to "us"}

\item{equation}{which equation to use - you can either set it to "us" for the
us 1978 equations, or "user supplied")}

\item{eq_df}{if you select "user supplied" for equation, this parameter should
be your equation dataframe, which (should have been reshaped by the
reshape_new_equation function prior)}

\item{beh}{lowercase string corresponding to the behavior term}

\item{which}{indicate whether you want the optimal behavior for "actor", "object",
or "both"}
}
\value{
3 digit EPA indicating the optimal behavior
}
\description{
Identify the behavior that would maximally confirm the identities of actor
and object pairing
}
