% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{merge_proteome}
\alias{merge_proteome}
\title{Add protein abundance to an \code{InteRactome}}
\usage{
merge_proteome(res)
}
\arguments{
\item{res}{an \code{InteRactome}}
}
\description{
Add protein abundance to an \code{InteRactome}. 
Protein abundance are obtained from CD4+ effector T cells.
}
