% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{merge_proteome}
\alias{merge_proteome}
\title{Add protein abundance to an \code{InteRactome}}
\usage{
merge_proteome(res, col_ID = "Protein.IDs", sep = c(";", "|", "-"))
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{col_ID}{name of \code{res} containing protein IDs}

\item{sep}{Set of separators used sequentially (from right to left) to 
identify distinct protein IDs}
}
\description{
Add protein abundance to an \code{InteRactome}. 
Protein abundance are obtained from CD4+ effector T cells.
}
