% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_identity.R
\name{modify_identity}
\alias{modify_identity}
\title{Modify Identity}
\usage{
modify_identity(
  d,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)
}
\arguments{
\item{d}{a dataframe from reshape events that should be only
actor_modifier and actor elements or object_modifier and object elements}

\item{equation_key}{a string corresponding to the equation key from actdata}

\item{equation_gender}{either average, male, or female, depending on if you are using gendered equations}
}
\value{
three digit EPA profile of the modified identity
}
\description{
A function that applies the modifier equations to find the EPA location for modifier
\itemize{
\item identity
}
}
\examples{
d <- tibble::tibble(actor_modifier = "tired", actor = "ceo")
d <- reshape_events_df(df = d, df_format = "wide", dictionary_key = "usfullsurveyor2015", dictionary_gender = "average")
tired_ceo <- modify_identity(d = d, equation_key = "us2010", equation_gender = "average")
}
