% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_identity.R
\name{modify_identity}
\alias{modify_identity}
\title{Modify Identity}
\usage{
modify_identity(
  data,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)
}
\arguments{
\item{id_info}{a dataframe from reshape events that should be only
actor_modifier and actor elements or object_modifier and object elements}

\item{eq_info}{the equation information you are using, should be in the
form {name}_{gender}}
}
\value{
three digit EPA profile of the modified identity
}
\description{
A function that applies the modifier equations to find the EPA location for modifier
\itemize{
\item identity
}
}
\examples{

}
