% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_behavior.R
\name{optimal_behavior}
\alias{optimal_behavior}
\title{Calculate the Optimal Behavior for the Actor following an Event}
\usage{
optimal_behavior(df, equation_info)
}
\arguments{
\item{df}{data that has been reshaped by the events_df}

\item{equation_info}{is a string that corresponds to "{equationkey}_{gender}"
from actdata}
}
\value{
3 digit EPA indicating the optimal behavior
}
\description{
Calculate the Optimal Behavior for the Actor following an Event
}
\examples{

opt_behavior_example <- tibble::tibble(actor = "teenager",
behavior = "beam_at",object = "friend")

opt_behavior_df <- reshape_events_df(df = opt_behavior_example,
df_format = "wide", dictionary_key = "indiana2003",
dictionary_gender = "male")

opt_b <- optimal_behavior(df = opt_behavior_df, equation_info = "nc1978_male")



}
