% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_behavior.R
\name{optimal_behavior}
\alias{optimal_behavior}
\title{Calculate the Optimal Behavior for the Actor following an Event}
\usage{
optimal_behavior(act, beh, obj, dictionary = "us")
}
\arguments{
\item{beh}{lowercase string corresponding to the behavior term}

\item{dictionary}{which dictionary to use, currently set to "us"}

\item{actor}{lowercase string corresponding to the actor identity}

\item{object}{lowercase string corresponding to the object identity}
}
\value{
3 digit EPA indicating the optimal behavior
}
\description{
Calculate the Optimal Behavior for the Actor following an Event
}
