% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{plot_2D_stoichio}
\alias{plot_2D_stoichio}
\title{Plot abundance versus interaction stoichiometries}
\usage{
plot_2D_stoichio(res, condition = "max", xlim = NULL, ylim = NULL,
  N_display = 30)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{condition}{condition selected. If "max", the maximum stoichiometry across conditions will be used.}

\item{xlim}{range of x values}

\item{ylim}{range of y values}

\item{N_display}{maximum number of protein to display}
}
\value{
a plot
}
\description{
Plot abundance versus interaction stoichiometries
}
