% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_2D_stoichio}
\alias{plot_2D_stoichio}
\title{Plot abundance versus interaction stoichiometries}
\usage{
plot_2D_stoichio(res, condition = "max", xlim = NULL, ylim = NULL,
  N_display = 30, only_interactors = FALSE,
  fill_values = c(not_regulated = "black", induced = "red", repressed =
  "blue", bait = "yellow"), color_values = c(not_regulated = "black",
  induced = "red", repressed = "blue", bait = "black"), shape = 21,
  stroke = 1, p_val_thresh = 0.05, fold_change_thresh = 1,
  ref_condition = res$conditions[1], label_size_min = 1,
  label_size_max = 3, label_size_scale_factor = 25,
  label_range = NULL, ...)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{condition}{condition selected. If "max", the maximum stoichiometry across conditions will be used.}

\item{xlim}{range of x values}

\item{ylim}{range of y values}

\item{N_display}{maximum number of protein to display}

\item{only_interactors}{display only interactors 
(identified using the function \code{identify_interactors()})}

\item{fill_values}{color vector passed to \code{scale_fill_manual()}}

\item{color_values}{color vector passed to \code{scale_color_manual()}}

\item{shape}{Point shape aesthetics passed to \code{geom_point()}}

\item{stroke}{Point stroke aesthetics passed to \code{geom_point()}}

\item{p_val_thresh}{Threshold on p-value used to identify regulated interactions}

\item{fold_change_thresh}{Threshold on fold-change used to identify regulated interactions}

\item{ref_condition}{Reference condition used to identify regulated interactions}

\item{label_size_min}{minimum label size (between 0 and \code{label_size_max})}

\item{label_size_max}{maximum label size (a threshold on log10(fold-change))}

\item{label_size_scale_factor}{scale label size according to plot range (the higher the bigger the label)}

\item{label_range}{if NULL, scales labels according to plot range (in log10 scale).}

\item{...}{parameters passed to \code{geom_text_repel()}}
}
\value{
a plot
}
\description{
Plot abundance versus interaction stoichiometries
}
