% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{plot_density}
\alias{plot_density}
\title{Plot points with density background with correlation coefficient}
\usage{
plot_density(df, var_x = names(df)[1], var_y = names(df)[2])
}
\arguments{
\item{df}{a data.frame}

\item{var_x}{name of the x variable}

\item{var_y}{name of the y variable}
}
\value{
a plot
}
\description{
Plot points with density background with correlation coefficient
}
