% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_indirect_interactions}
\alias{plot_indirect_interactions}
\title{Plot indirect interactions}
\usage{
plot_indirect_interactions(score,
  var_threshold = "max_delta_stoichio_log", threshold = 1,
  min_range = NULL, save_dir = NULL, plot_width = 2,
  plot_height = 2, show_legend = FALSE)
}
\arguments{
\item{score}{output of the \code{identify_indirect_interactions} function}

\item{var_threshold}{Variable of \code{score} on which the threshold will be applied}

\item{threshold}{maximum difference between observed and predicted stoichiometries (in log10)}

\item{min_range}{Minimum range (in log10 scale) displayed on the y-axis (centered on the mean)}

\item{save_dir}{path to the directory where the plot will be saved}

\item{plot_width}{set plot width}

\item{plot_height}{set plot height}

\item{show_legend}{logical, shows plot legend}
}
\description{
Plot indirect interactions
}
