% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{plot_indirect_interactions}
\alias{plot_indirect_interactions}
\title{Plot indirect interactions}
\usage{
plot_indirect_interactions(score, threshold = 1, save_file = NULL,
  facet = FALSE, plot_width = NULL, plot_height = NULL)
}
\arguments{
\item{score}{output of the \code{identify_indirect_interactions} function}

\item{threshold}{maximum difference between observed and predicted stoichiometries (in log10)}

\item{save_file}{path were the plot will be saved}
}
\description{
Plot indirect interactions
}
