% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_QC}
\alias{plot_QC}
\title{Quality check plots for preprocessed AP-MS data}
\usage{
plot_QC(data)
}
\arguments{
\item{data}{an \code{Interactome} or preprocessed data as obtained using function \code{preprocess_data()}}
}
\value{
Several QC plots
}
\description{
Quality check plots for preprocessed AP-MS data
}
