% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{plot_volcanos}
\alias{plot_volcanos}
\title{Plot protein enrichement fold-change versus p-value}
\usage{
plot_volcanos(res, labels = NULL, N_print = 15, conditions = NULL,
  p_val_thresh = 0.05, fold_change_thresh = 2, x0 = NULL, c = NULL,
  save_file = NULL, xlim = NULL, ylim = NULL,
  asinh_transform = TRUE, norm = FALSE)
}
\arguments{
\item{res}{an \code{InteRactome}}

\item{labels}{labels for proteins in plot. Must the same length as \code{res$names}}

\item{N_print}{maximum of protein labels to display}

\item{conditions}{conditions to plot}

\item{p_val_thresh}{threshold on p-value to display}

\item{fold_change_thresh}{threshold on fold-change to display}

\item{save_file}{path of output file (.pdf)}

\item{xlim}{range of x values}

\item{ylim}{range of y values}

\item{asinh_transform}{logical, display asinh(log10(p-value)) on the y-axis}

\item{norm}{Use normalized fold-changes}
}
\value{
a plot
}
\description{
Plot protein enrichement fold-change versus p-value
}
