% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify_actor.R
\name{reidentify_actor}
\alias{reidentify_actor}
\title{Function to Generate the Re-identification of the Actor}
\usage{
reidentify_actor(
  act,
  beh,
  obj,
  dictionary_key,
  gender,
  equation_key,
  eq_df = NULL
)
}
\arguments{
\item{act}{lowercase string corresponding to the actor identity}

\item{beh}{lowercase string corresponding to the behavior term}

\item{obj}{lowercase string corresponding to the object identity}

\item{dictionary_key}{a string corresponding to the dictionary from actdata you are using for cultural EPA measurements}

\item{gender}{either average, male, or female, depending on if you are using gendered equations}

\item{equation_key}{a string corresponding to the equation key from actdata}

\item{eq_df}{if you select "user supplied" for equation, this parameter should
be your equation dataframe, which (should have been reshaped by the
reshape_new_equation function prior)}
}
\value{
dataframe with 3 columns corresponding to the EPA of optimal actor identity relabel
}
\description{
Function to Generate the Re-identification of the Actor
}
\examples{
reidentify_actor(act = "ceo", beh = "advise", obj = "benefactor", dictionary_key = "usfullsurveyor2015",
gender = "average", equation_key = "us2010")
}
