% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify_object.R
\name{reidentify_object}
\alias{reidentify_object}
\title{Function to Generate the Re-identification of the Object}
\usage{
reidentify_object(
  data,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)
}
\arguments{
\item{df}{data that has been reshaped by the events_df}

\item{equation_info}{is a string that corresponds to "{equationkey}_{gender}" from actdata}
}
\value{
dataframe with 3 columns corresponding to the EPA of optimal object identity relabel
}
\description{
Function to Generate the Re-identification of the Object
}
\examples{
d <- tibble::tibble(actor = "ceo", behavior = "advise", object = "benefactor")
d <- reshape_events_df(df = d, df_format = "wide", dictionary_key = "usfullsurveyor2015", dictionary_gender = "average")
reidentify_object(data = d, equation_key= "us2010", equation_gender = "average")
}
