% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reidentify_object.R
\name{reidentify_object}
\alias{reidentify_object}
\title{Function to Generate the Re-identification of the Object}
\usage{
reidentify_object(act, beh, obj, dictionary = "us")
}
\arguments{
\item{act}{lowercase string corresponding to the actor identity}

\item{beh}{lowercase string corresponding to the behavior term}

\item{obj}{lowercase string corresponding to the object identity}

\item{dictionary}{which dictionary to use, currently set to "us"}
}
\value{
dataframe with 3 columns corresponding to the EPA of optimal object identity relabel
}
\description{
Function to Generate the Re-identification of the Object
}
\examples{
reidentify_object(act = "brute", beh = "work", obj = "cook")
}
