% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{row_mean}
\alias{row_mean}
\title{Compute the mean by row}
\usage{
row_mean(df, na.rm = TRUE, log = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{na.rm}{logical, remove NA values}

\item{log}{logical, use geometric mean instead of arithmetic mean}
}
\value{
A numeric vector
}
\description{
Compute the mean by row
}
