% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRact.R
\name{smooth.InteRactome}
\alias{smooth.InteRactome}
\title{Smooth selected variables across conditions}
\usage{
\method{smooth}{InteRactome}(res, n = 1, order_conditions = NULL,
  var_smooth = c("fold_change", "p_val"))
}
\description{
Smooth selected variables across conditions
}
