% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient_impression.R
\name{transient_impression}
\alias{transient_impression}
\title{Calculate the Transient Impression after an Event}
\usage{
transient_impression(
  act,
  beh,
  obj,
  dictionary_key,
  gender,
  equation_key,
  eq_df = NULL
)
}
\arguments{
\item{beh}{lowercase string corresponding to the behavior term}

\item{dictionary_key}{a string corresponding to the dictionary from actdata you are using for cultural EPA measurements}

\item{gender}{either average, male, or female, depending on if you are using gendered equations}

\item{equation_key}{a string corresponding to the equation key from actdata}

\item{eq_df}{if you select "user supplied" for equation key, this parameter should
be your equation dataframe, which (should have been reshaped by the
reshape_new_equation function prior)}

\item{actor}{lowercase string corresponding to the actor identity}

\item{object}{lowercase string corresponding to the object identity}
}
\value{
dataframe in long format, with one row for each element-dimension of the event, columns for fundamental sentiment and transient impression.
}
\description{
Calculate the Transient Impression after an Event
}
\examples{
transient_impression(act = "ceo", beh = "advise", obj = "benefactor", dictionary_key = "usfullsurveyor2015",
gender = "average", equation_key = "us2010")

}
