% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient_impression.R
\name{transient_impression}
\alias{transient_impression}
\title{Calculate the Transient Impression after an Event}
\usage{
transient_impression(act, beh, obj, dictionary = "us")
}
\arguments{
\item{beh}{lowercase string corresponding to the behavior term}

\item{dictionary}{which dictionary to use, currently set to "us"}

\item{actor}{lowercase string corresponding to the actor identity}

\item{object}{lowercase string corresponding to the object identity}
}
\value{
dataframe in long format, with one row for each element-dimension of the event, columns for fundamental sentiment and transient impression.
}
\description{
Calculate the Transient Impression after an Event
}
\examples{
transient_impression("ceo", "advise", "benefactor")
}
