% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient_impression.R
\name{transient_impression}
\alias{transient_impression}
\title{Calculate the Transient Impression after an Event}
\usage{
transient_impression(df, equation_info, eq_df = NULL, ...)
}
\arguments{
\item{df}{data that has been reshaped by the events_df}

\item{equation_info}{is a string that corresponds to "{equationkey}_{gender}" from actdata}
}
\value{
dataframe in long format, with one row for each element-dimension of the event, columns for fundamental sentiment and transient impression.
}
\description{
Calculate the Transient Impression after an Event
}
\examples{
d <- tibble::tibble(actor_modifier = "tired", actor = "ceo", behavior = "advise", object = "benefactor")
d <- reshape_events_df(df = d, df_format = "wide", dictionary_key = "usfullsurveyor2015", dictionary_gender = "average")
transient_impression(df = d, equation_info = "us2010_average")

}
